Source code/requirements.txt
pandas
scipy
scikit-learn
scikit-cmeans
pyclustering

rpy2
python-weka-wrapper3
